refactor dropzone modal into smaller comps
